requirements.txt
django
djangorestframework
django-filter
django-environ
drf-yasg